generate_mri_fastapi/pyproject.toml
[tool.poetry]
name = "generate mri fastapi"
version = "0.1.0"
description = ""
authors = ["hanahh080601 <hanlehngoc080601@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
nibabel = "5.1.0"
numpy = "1.23.5"
Pillow = "9.5.0"
torch = "2.0.0"
torchio = "0.18.90"
torchvision = "0.15.1"
tqdm = "4.65.0"
pydantic = "^1.10.8"
fastapi = "^0.95.2"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
pymongo = "^4.3.3"
python-multipart = "^0.0.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
